updated code of file upload on cloudinary

+++ src/utils/cloudinary.js
@@ -1,27 +1,25 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import dotenv from 'dotenv';
+dotenv.config();
 
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
     api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret: process.env.CLOUDINARY_API_SECRET
+    api_secret: process.env.CLOUDINARY_API_SECRET,
 });
-
 const uploadOnCloudnary = async (localPathFile) => {
     try {
         if (!localPathFile) return null;
-        const response = await cloudinary.uploader.upload(localPathFile, resourse_type = "auto", (error, result) => {
-            if (error) {
-                console.log(error);
-                return null;
-            }
-            if (!result) return null;
-            if (result) return result.url;
-        })
-        console.log(response.url)
-        return response
+        const response = await cloudinary.uploader.upload(localPathFile, { resource_type: "auto" })
+        console.log("file is uploaded to cloudinary", response.url)
+        if (fs.existsSync(localPathFile)) { fs.unlinkSync(localPathFile)
+            return response
+         }
+         
     } catch (error) {
-        fs.unlinkSync(localPathFile)
+        if (fs.existsSync(localPathFile)) { fs.unlinkSync(localPathFile) }
+        console.log("error in upload on cloudinary", error)
         return null
     }
 }
